--- main-library/src/main/java/rocket/agile/com/mainlibrary/Interface/NetworkingListener.java
@@ -1,10 +1,0 @@
-package rocket.agile.com.mainlibrary.Interface;
-
-/**
- * Created by KeithK on 6/11/17.
- */
-
-public interface NetworkingListener {
-
-    void onNetworkingListener(int responseCode, String result);
-}
