+++ main-library/src/main/java/rocket/agile/com/mainlibrary/realm/RealmPersistence.java
@@ -2,12 +2,12 @@ package rocket.agile.com.mainlibrary.realm;
 
 import io.realm.Realm;
 import io.realm.RealmConfiguration;
-import rocket.agile.com.mainlibrary.actionItems.ActionPhone;
+import rocket.agile.com.mainlibrary.model.actionItems.ActionPhone;
 import rocket.agile.com.mainlibrary.activity.MasterView;
-import rocket.agile.com.mainlibrary.actionItems.ActionEmail;
-import rocket.agile.com.mainlibrary.actionItems.ActionList;
+import rocket.agile.com.mainlibrary.model.actionItems.ActionEmail;
+import rocket.agile.com.mainlibrary.model.actionItems.ActionList;
 import rocket.agile.com.mainlibrary.model.DataManager;
-import rocket.agile.com.mainlibrary.model.Values;
+import rocket.agile.com.mainlibrary.model.actionItems.Values;
 
 /**
  * Created by keithkowalski on 4/21/17.
